Fixed clippy warnings from Rust 1.66.0

+++ mithril-stm/benches/size_benches.rs
@@ -57,7 +57,7 @@ where
         k,
         m,
         nparties,
-        aggr.to_bytes().len() as usize,
+        aggr.to_bytes().len(),
     );
 }
